# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Production ready API template using Echo",
        "title": "Go Backend Boilerplate API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/api/v1/auth/signup": {
            "post": {
                "description": "Create asd new user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "v1/auth"
                ],
                "summary": "Create user",
                "parameters": [
                    {
                        "description": "Signup",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.PasswordSignUpPayload"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/auth.PasswordSignUpPayload"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errs.HTTPError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "auth.PasswordSignUpPayload": {
            "type": "object",
            "required": [
                "name",
                "password",
                "username"
            ],
            "properties": {
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "minLength": 6
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "errs.Action": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
        "errs.FieldError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "field": {
                    "type": "string"
                }
            }
        },
        "errs.HTTPError": {
            "type": "object",
            "properties": {
                "action": {
                    "$ref": "#/definitions/errs.Action"
                },
                "code": {
                    "type": "string"
                },
                "errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/errs.FieldError"
                    }
                },
                "message": {
                    "type": "string"
                },
                "override": {
                    "type": "boolean"
                },
                "status": {
                    "type": "integer"
                }
            }
        }
    }
}